+++ HISTORY.md
@@ -68,6 +68,7 @@
 ### 2022-11-28
 * objects: new recordUserEdit() mechanism
 * objects: migrated objects to new recordUserEdit mechanism
+* blocks: migrated blocks to new recordUserEdit mechanism
 
 ### 2022-11-26
 * threads: order "my scripts" list by their top-down position in the scripting area